backend/requirements.txt
fastapi
matplotlib
pandas
seaborn
uvicorn
requests==2.31.0
httpx
keyboard
pillow
mss
opencv-python 
easyocr
fuzzywuzzy
python-Levenshtein
torch
transformers